external/periodic_shapes/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
configparser==4.0.2
cycler==0.10.0
docker-pycreds==0.4.0
gitdb==4.0.2
GitPython==3.1.0
gql==0.2.0
graphql-core==1.1
idna==2.9
kiwisolver==1.1.0
matplotlib==3.1.3
numpy==1.18.1
nvidia-ml-py3==7.352.0
pathtools==0.1.2
Pillow==6.2.2
plotly==4.5.2
promise==2.3
psutil==5.7.0
pyparsing==2.4.6
python-dateutil==2.8.1
python-dotenv==0.12.0
PyYAML==5.3
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
sentry-sdk==0.14.2
shortuuid==0.5.0
six==1.14.0
smmap==3.0.1
subprocess32==3.5.4
torch==1.4.0
torchvision==0.5.0
urllib3==1.25.8
wandb==0.8.28
watchdog==0.10.2
pylint
yapf
nbformat==5.0.4	
pykeops==1.3
